webscraper/requirements.txt
selenium==4.16.0
beautifulsoup4==4.10.0
webdriver-manager==4.0.1
motor==3.4.0
python-decouple==3.6